Tided up code for PR
