+++ src/CommitmentsV2/SFA.DAS.CommitmentsV2.MessageHandlers/DependencyResolution/IoC.cs
@@ -8,7 +8,6 @@ namespace SFA.DAS.CommitmentsV2.MessageHandlers.DependencyResolution
         public static void Initialize(Registry registry)
         {
             registry.IncludeRegistry<NServiceBusUnitOfWorkRegistry>();
-
             registry.IncludeRegistry<DefaultRegistry>();
         }
     }
