+++ src/CommitmentsV2/SFA.DAS.CommitmentsV2.MessageHandlers/NServiceBus/EndpointConfigurationExtensions.cs
@@ -13,5 +13,4 @@ namespace SFA.DAS.CommitmentsV2.MessageHandlers.NServiceBus
             return config;
         }
     }
-
 }
