+++ src/CommitmentsV2/SFA.DAS.CommitmentsV2.MessageHandlers/SFA.DAS.CommitmentsV2.MessageHandlers.csproj
@@ -37,9 +37,6 @@
     <ProjectReference Include="..\SFA.DAS.CommitmentsV2\SFA.DAS.CommitmentsV2.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Configuration\" />
-  </ItemGroup>
-  <ItemGroup>
     <None Update="appsettings.Development.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
