--- src/CommitmentsV2/SFA.DAS.CommitmentsV2/Test/Class1.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SFA.DAS.CommitmentsV2.Test
-{
-    public class Class1 : Interface1
-    {
-        public string Simple { get; set; }
-    }
-}
