--- src/CommitmentsV2/SFA.DAS.CommitmentsV2/Test/Interface1.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SFA.DAS.CommitmentsV2.Test
-{
-    public interface Interface1
-    {
-        string Simple { get; set; }
-    }
-}
